requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
requests
streamlit
python-dotenv
